Function await::examples::async::any

pub fn any<T: Send + 'static>(xs: Vec<Box<AwaitBox<T> + Send>>) -> Box<AwaitBox<T>>

Returns an Await that will await for any of the inputs to finish. This method returns immediatly.

TODO: once impl Trait lands we can change this to .. -> impl Await<T>