Module await::examples::promise

"Crate" for rerpresenting the async abstraction of promises / futures

A Future represents the potential of a value in the future

A Promise is used to complete (a.k.a fulfill) the Future.

One can say that a Future<T> is in essence an Await<T>, and they wouldn't be wrong.

Example: This does not need to be on this crate

Structs

Future
At some point in time, it will hold a value T

Promise
Used to fulfill a promise (in other words, complete the Future)

PromiseDroppedError
Signals that the Promise dropped before being fullfilled. Liar...