Trait await::Await

pub trait Await<T> {
    fn await(self) -> T;
}

Represents then notion of something that we could have to await for.

The keyword await could be added to the language.

Required Methods

fn await(self) -> T

It will return a value when it's ready