Trait await::AwaitBox

pub trait AwaitBox<T> {
    fn await_box(self: Box<Self>) -> T;
}

AwaitBox is a version of the Await intended for use with boxed objects.

The idea is that where one would normally store a Box<Await<T>> in a data structure, you should use Box<AwaitBox<T>>. The two traits behave essentially the same, except that a AwaitBox can only be called if it is boxed.

Note that AwaitBox may be deprecated in the future if Box<Await<T>> become directly usable.

Required Methods

fn await_box(self: Box<Self>) -> T